Major overhaul of CyDelete: - No longer bolt ourselves onto every SBIcon instance, and instead add the proper messages to the SBApplicationIcon subclass, as SBApplicationIcon represents system applications. - Uninstalls are now done in an NSOperationQueue in the background. - Uninstalls are now done by hooking and calling [SBApplicationController uninstallApplication:...] - Non-cydia deletion is no longer optional. It is always on.
This is going to be version 2.0. No more progress HUDs.

Accidentally fixed the bug where (X)s would be added to webclips installed by provisioning profiles.
This overhaul allows for greater flexibility when being used with systems that do not use SBIcons (PogoPlank) and when SBIcon subclasses are made by other developers (STS for Stack!)
This also fixes the issue where I'm creating my own UIAlertViews over top of SpringBoard. Implementing the uninstallAlert* messages in SBApplicationIcon allows SpringBoard to handle asking for deletion instead of me doing it myself.
When an icon's (X) is tapped it's package is stored in a CFDictionary (keyed on the SBIcon instance) for safekeeping.

 Hook.mm                                                                        |  451 ++++------
 layout/DEBIAN/control                                                          |    2 
 layout/System/Library/PreferenceBundles/CyDeleteSettings.bundle/CyDelete.plist |   28 
3 files changed, 212 insertions(+), 269 deletions(-)